add(print.c, tree.c) separate func
