+++ playground/ta/parser_test/test.h
@@ -45,23 +45,30 @@ struct s_tree
 	t_tree		*rhs;
 };
 
-// token
-void	print_token(t_token *token_head);
-void	print_token_node(t_token *token);
+// tokenizer
 void	free_token(t_token **token);
 t_token	*tokenize(char **split);
 
-// tree
-t_tree	*new_node(t_node_kind kind, t_tree *lhs, t_tree *rhs);
-t_tree	*new_num_leaf(int val);
+// parser
 
 t_tree	*expression(t_token **token);
 t_tree	*term(t_token **token);
 t_tree	*primary(t_token **token);
 
 
-void	print_tree(t_tree *tree);
+// tree
+t_tree	*new_node(t_node_kind kind, t_tree *lhs, t_tree *rhs);
+t_tree	*new_num_leaf(int val);
 void	free_tree(t_tree **node);
 char	get_operator_char(t_tree *node);
+size_t	get_depth(t_tree *node);
+
+
+// print
+void	print_split(char **split);
+void	print_token(t_token *token_head);
+void	print_token_node(t_token *token);
+void	print_tree(t_tree *tree);
+
 
 #endif //TEST_H
